redo of #47202 (#47211)
* redo of #47202

* fix nested test module references
