+++ types/function-bind/function-bind-tests.ts
@@ -1,5 +1,7 @@
 import bind = require('function-bind');
 
+declare function call<T, A extends any[], R>(this: (this: T, ...args: A) => R, thisArg: T, ...args: A): R;
+
 declare const string: string;
 declare const number: number;
 declare const boolean: boolean;
@@ -11,35 +13,45 @@ declare const boolean: boolean;
 declare function expectType<T>(value: T): T;
 
 // $ExpectType (thisArg: any, start?: number | undefined, end?: number | undefined) => any[]
-const slice = expectType<(thisArg: any, start?: number, end?: number) => any[]>(
-    bind.call(Function.call, Array.prototype.slice),
-);
+const slice = expectType<(thisArg: any, start?: number, end?: number) => any[]>(bind.call(call, Array.prototype.slice));
 
 // $ExpectType (start?: number | undefined, end?: number | undefined) => any[]
-expectType<(start?: number, end?: number) => any[]>(bind.call(Function.call, Array.prototype.slice, null));
+expectType<(start?: number, end?: number) => any[]>(bind.call(call, Array.prototype.slice, null));
 
 // $ExpectType (end?: number | undefined) => any[]
-expectType<(end?: number) => any[]>(bind.call(Function.call, Array.prototype.slice, ['a'], 1));
+expectType<(end?: number) => any[]>(bind.call(call, Array.prototype.slice, ['a'], 1));
 
 slice(['a']);
 
-// $ExpectType (...args: string[]) => boolean
-expectType<(...args: string[]) => boolean>(bind.call(Boolean, null, String(), '2', '3', '4', '5'));
+// $ExpectType () => boolean
+bind.call(
+    Boolean,
+    null,
+    '1',
+    // @ts-expect-error
+    '2',
+    '3',
+    '4',
+    '5',
+);
 
-// $ExpectType (...args: string[]) => boolean
-expectType<(...args: string[]) => boolean>(bind.apply(Boolean, [null, '1', '2', '3', '4', '5']));
+// $ExpectType (value?: unknown) => boolean
+bind.call(Boolean, null);
+
+// $ExpectType () => boolean
+expectType<() => boolean>(bind.apply(Boolean, [null, '1', '2', '3', '4', '5']));
 
 // Class compatibility:
 class Foo {
     constructor(public string: string, public number: number) {}
-    static bind: typeof bind;
+    declare static bind: typeof bind;
 }
 
 // bind.call():
 // $ExpectType new (string: string, number: number) => Foo
 bind.call(Foo, null);
 
-// $ExpectType new (string: string, number: number) => Foo
+// $ExpectType typeof Foo
 Foo.bind(null);
 
 // $ExpectType new (number: number) => Foo
@@ -54,16 +66,16 @@ bind.call(Foo, null, string, number);
 // $ExpectType new () => Foo
 Foo.bind(null, string, number);
 
-// $ExpectType new () => Foo
+// @ts-expect-error
 bind.call(Foo, null, string, number, boolean);
 
-// $ExpectType new () => Foo
+// @ts-expect-error
 Foo.bind(null, string, number, boolean);
 
-// $ExpectType new () => Foo
+// @ts-expect-error
 bind.call(Foo, null, string, number, boolean, undefined);
 
-// $ExpectType new () => Foo
+// @ts-expect-error
 Foo.bind(null, string, number, boolean, undefined);
 
 // bind.apply():
