+++ types/function-bind/implementation.d.ts
@@ -1,3 +1,5 @@
+export = bind;
+
 //#region bind():
 /**
  * For a given function, creates a bound function that has the same body as the original function.
@@ -5,179 +7,72 @@
  * @param thisArg The object to be used as the this object.
  * @param args Arguments to bind to the parameters of the function.
  */
-declare function bind<T, A extends any[], R>(this: (this: T, ...args: A) => R, thisArg: T): (...args: A) => R;
-declare function bind<T, A0, A extends any[], R>(
-    this: (this: T, arg0: A0, ...args: A) => R,
-    thisArg: T,
-    arg0: A0,
-): (...args: A) => R;
-declare function bind<T, A0, A1, A extends any[], R>(
-    this: (this: T, arg0: A0, arg1: A1, ...args: A) => R,
-    thisArg: T,
-    arg0: A0,
-    arg1: A1,
-): (...args: A) => R;
-declare function bind<T, A0, A1, A2, A extends any[], R>(
-    this: (this: T, arg0: A0, arg1: A1, arg2: A2, ...args: A) => R,
-    thisArg: T,
-    arg0: A0,
-    arg1: A1,
-    arg2: A2,
-): (...args: A) => R;
-declare function bind<T, A0, A1, A2, A3, A extends any[], R>(
-    this: (this: T, arg0: A0, arg1: A1, arg2: A2, arg3: A3, ...args: A) => R,
-    thisArg: T,
-    arg0: A0,
-    arg1: A1,
-    arg2: A2,
-    arg3: A3,
-): (...args: A) => R;
-declare function bind<T, AX, R>(this: (this: T, ...args: AX[]) => R, thisArg: T, ...args: AX[]): (...args: AX[]) => R;
+declare function bind<T>(this: T, thisArg: ThisParameterType<T>): OmitThisParameter<T>;
 
-declare function bind<A extends any[], R>(this: new (...args: A) => R, thisArg: any): new (...args: A) => R;
-declare function bind<A0, A extends any[], R>(
-    this: new (arg0: A0, ...args: A) => R,
-    thisArg: any,
-    arg0: A0,
+// CallableFunction:
+declare function bind<T, AX extends readonly unknown[], A extends readonly unknown[], R>(
+    this: (this: T, ...args: [...bound: AX, ...args: A]) => R,
+    thisArg: T,
+    ...bound: AX
+): (...args: A) => R;
+
+// NewableFunction:
+declare function bind<AX extends readonly unknown[], A extends readonly unknown[], R>(
+    this: new (...args: [...bound: AX, ...args: A]) => R,
+    thisArg: unknown,
+    ...bound: AX
 ): new (...args: A) => R;
-declare function bind<A0, A1, A extends any[], R>(
-    this: new (arg0: A0, arg1: A1, ...args: A) => R,
-    thisArg: any,
-    arg0: A0,
-    arg1: A1,
-): new (...args: A) => R;
-declare function bind<A0, A1, A2, A extends any[], R>(
-    this: new (arg0: A0, arg1: A1, arg2: A2, ...args: A) => R,
-    thisArg: any,
-    arg0: A0,
-    arg1: A1,
-    arg2: A2,
-): new (...args: A) => R;
-declare function bind<A0, A1, A2, A3, A extends any[], R>(
-    this: new (arg0: A0, arg1: A1, arg2: A2, arg3: A3, ...args: A) => R,
-    thisArg: any,
-    arg0: A0,
-    arg1: A1,
-    arg2: A2,
-    arg3: A3,
-): new (...args: A) => R;
-declare function bind<AX, R>(this: new (...args: AX[]) => R, thisArg: any, ...args: AX[]): new (...args: AX[]) => R;
 //#endregion
 
 declare namespace bind {
     //#region bind.call():
-    /**
-     * Creates a bound function with the specified object as the this value and the specified rest arguments as the arguments.
-     * @param thisArg The object to be used as the this object.
-     * @param args Argument values to be passed to the function.
-     */
     // CallableFunction:
-    function call<T, A extends any[], R>(func: (this: T, ...args: A) => R, thisArg: T): (...args: A) => R;
-    function call<T, A0, A extends any[], R>(
-        func: (this: T, arg0: A0, ...args: A) => R,
+    function call<T, AX extends readonly unknown[], A extends readonly unknown[], R>(
+        func: (this: T, ...args: [...bound: AX, ...args: A]) => R,
         thisArg: T,
-        arg0: A0,
+        ...bound: AX
     ): (...args: A) => R;
-    function call<T, A0, A1, A extends any[], R>(
-        func: (this: T, arg0: A0, arg1: A1, ...args: A) => R,
-        thisArg: T,
-        arg0: A0,
-        arg1: A1,
-    ): (...args: A) => R;
-    function call<T, A0, A1, A2, A extends any[], R>(
-        func: (this: T, arg0: A0, arg1: A1, arg2: A2, ...args: A) => R,
-        thisArg: T,
-        arg0: A0,
-        arg1: A1,
-        arg2: A2,
-    ): (...args: A) => R;
-    function call<T, A0, A1, A2, A3, A extends any[], R>(
-        func: (this: T, arg0: A0, arg1: A1, arg2: A2, arg3: A3, ...args: A) => R,
-        thisArg: T,
-        arg0: A0,
-        arg1: A1,
-        arg2: A2,
-        arg3: A3,
-    ): (...args: A) => R;
-    function call<T, AX, R>(func: (this: T, ...args: AX[]) => R, thisArg: T, ...args: AX[]): (...args: AX[]) => R;
 
     // NewableFunction:
-    function call<A extends any[], R>(func: new (...args: A) => R, thisArg: unknown): new (...args: A) => R;
-    function call<A0, A extends any[], R>(
-        func: new (arg0: A0, ...args: A) => R,
+    function call<AX extends readonly unknown[], A extends readonly unknown[], R>(
+        func: new (...args: [...bound: AX, ...args: A]) => R,
         thisArg: unknown,
-        arg0: A0,
+        ...bound: AX
     ): new (...args: A) => R;
-    function call<A0, A1, A extends any[], R>(
-        func: new (arg0: A0, arg1: A1, ...args: A) => R,
-        thisArg: unknown,
-        arg0: A0,
-        arg1: A1,
-    ): new (...args: A) => R;
-    function call<A0, A1, A2, A extends any[], R>(
-        func: new (arg0: A0, arg1: A1, arg2: A2, ...args: A) => R,
-        thisArg: unknown,
-        arg0: A0,
-        arg1: A1,
-        arg2: A2,
-    ): new (...args: A) => R;
-    function call<A0, A1, A2, A3, A extends any[], R>(
-        func: new (arg0: A0, arg1: A1, arg2: A2, arg3: A3, ...args: A) => R,
-        thisArg: unknown,
-        arg0: A0,
-        arg1: A1,
-        arg2: A2,
-        arg3: A3,
-    ): new (...args: A) => R;
-    function call<AX, R>(func: new (...args: AX[]) => R, thisArg: unknown, ...args: AX[]): new (...args: AX[]) => R;
     //#endregion
 
     //#region bind.apply():
-    /**
-     * Creates a bound function with the specified object as the this value and the elements of specified array as the arguments.
-     * @param thisArg The object to be used as the this object.
-     * @param args An array of argument values to be passed to the function.
-     */
     // CallableFunction:
-    function apply<T, A extends any[], R>(func: (this: T, ...args: A) => R, args: [T]): (...args: A) => R;
-    function apply<T, A0, A extends any[], R>(
-        func: (this: T, arg0: A0, ...args: A) => R,
-        args: [T, A0],
+    function apply<T, AX extends readonly unknown[], A extends readonly unknown[], R>(
+        func: (this: T, ...args: [...bound: AX, ...args: A]) => R,
+        args: readonly [thisArg: T, ...bound: AX],
     ): (...args: A) => R;
-    function apply<T, A0, A1, A extends any[], R>(
-        func: (this: T, arg0: A0, arg1: A1, ...args: A) => R,
-        args: [T, A0, A1],
-    ): (...args: A) => R;
-    function apply<T, A0, A1, A2, A extends any[], R>(
-        func: (this: T, arg0: A0, arg1: A1, arg2: A2, ...args: A) => R,
-        args: [T, A0, A1, A2],
-    ): (...args: A) => R;
-    function apply<T, A0, A1, A2, A3, A extends any[], R>(
-        func: (this: T, arg0: A0, arg1: A1, arg2: A2, arg3: A3, ...args: A) => R,
-        args: [T, A0, A1, A2, A3],
-    ): (...args: A) => R;
-    function apply<T, AX, R>(func: (this: T, ...args: AX[]) => R, args: [T, ...AX[]]): (...args: AX[]) => R;
 
     // NewableFunction:
-    function apply<A extends any[], R>(func: new (...args: A) => R, args: [unknown]): new (...args: A) => R;
-    function apply<A0, A extends any[], R>(
-        func: new (arg0: A0, ...args: A) => R,
-        args: [unknown, A0],
+    // TODO: Figure out why this is necessary:
+    function apply<A extends readonly unknown[], R>(
+        func: new (...args: A) => R,
+        args: readonly [thisArg: unknown]
     ): new (...args: A) => R;
-    function apply<A0, A1, A extends any[], R>(
-        func: new (arg0: A0, arg1: A1, ...args: A) => R,
-        args: [unknown, A0, A1],
+    function apply<A0, A extends readonly unknown[], R>(
+        func: new (bound_0: A0, ...args: A) => R,
+        args: readonly [thisArg: unknown, bound_0: A0],
     ): new (...args: A) => R;
-    function apply<A0, A1, A2, A extends any[], R>(
-        func: new (arg0: A0, arg1: A1, arg2: A2, ...args: A) => R,
-        args: [unknown, A0, A1, A2],
+    function apply<A0, A1, A extends readonly unknown[], R>(
+        func: new (bound_0: A0, bound_1: A1, ...args: A) => R,
+        args: readonly [thisArg: unknown, bound_0: A0, bound_1: A1],
     ): new (...args: A) => R;
-    function apply<A0, A1, A2, A3, A extends any[], R>(
-        func: new (arg0: A0, arg1: A1, arg2: A2, arg3: A3, ...args: A) => R,
-        args: [unknown, A0, A1, A2, A3],
+    function apply<A0, A1, A2, A extends readonly unknown[], R>(
+        func: new (bound_0: A0, bound_1: A1, bound_2: A2, ...args: A) => R,
+        args: readonly [thisArg: unknown, bound_0: A0, bound_1: A1, bound_2: A2],
     ): new (...args: A) => R;
-    function apply<AX, R>(func: new (...args: AX[]) => R, args: [unknown, ...AX[]]): new (...args: AX[]) => R;
+    function apply<A0, A1, A2, A3, A extends readonly unknown[], R>(
+        func: new (bound_0: A0, bound_1: A1, bound_2: A2, bound_3: A3, ...args: A) => R,
+        args: readonly [thisArg: unknown, bound_0: A0, bound_1: A1, bound_2: A2, bound_3: A3],
+    ): new (...args: A) => R;
+    function apply<AX extends readonly unknown[], A extends readonly unknown[], R>(
+        func: new (...args: [...bound: AX, ...args: A]) => R,
+        args: readonly [thisArg: unknown, ...bound: AX],
+    ): new (...args: A) => R;
     //#endregion
 }
-
-export = bind;
