+++ types/function-bind/package.json
@@ -2,11 +2,11 @@
     "private": true,
     "types": "index",
     "typesVersions": {
-        ">=4.0.0-0": {
-            "*": ["ts4.0/*"]
+        "<=3.9": {
+            "*": ["ts3.9/*"]
         },
-        ">=3.3.0-0": {
-            "*": ["ts3.3/*"]
+        "<=3.2": {
+            "*": ["ts3.2/*"]
         }
     }
 }
