+++ types/function-bind/ts3.2/index.d.ts
