+++ types/function-bind/ts3.2/tsconfig.json
