+++ types/function-bind/ts3.2/tslint.json
