+++ types/function-bind/ts3.9/function-bind-tests.ts
@@ -1,6 +1,4 @@
-import bind = require('function-bind');
-
-declare function call<T, A extends any[], R>(this: (this: T, ...args: A) => R, thisArg: T, ...args: A): R;
+import bind = require('function-bind/ts3.9');
 
 declare const string: string;
 declare const number: number;
@@ -13,38 +11,28 @@ declare const boolean: boolean;
 declare function expectType<T>(value: T): T;
 
 // $ExpectType (thisArg: any, start?: number | undefined, end?: number | undefined) => any[]
-const slice = expectType<(thisArg: any, start?: number, end?: number) => any[]>(bind.call(call, Array.prototype.slice));
+const slice = expectType<(thisArg: any, start?: number, end?: number) => any[]>(
+    bind.call(Function.call, Array.prototype.slice),
+);
 
 // $ExpectType (start?: number | undefined, end?: number | undefined) => any[]
-expectType<(start?: number, end?: number) => any[]>(bind.call(call, Array.prototype.slice, null));
+expectType<(start?: number, end?: number) => any[]>(bind.call(Function.call, Array.prototype.slice, null));
 
 // $ExpectType (end?: number | undefined) => any[]
-expectType<(end?: number) => any[]>(bind.call(call, Array.prototype.slice, ['a'], 1));
+expectType<(end?: number) => any[]>(bind.call(Function.call, Array.prototype.slice, ['a'], 1));
 
 slice(['a']);
 
-// $ExpectType () => boolean
-bind.call(
-    Boolean,
-    null,
-    '1',
-    // @ts-expect-error
-    '2',
-    '3',
-    '4',
-    '5',
-);
+// $ExpectType (...args: string[]) => boolean
+bind.call(Boolean, null, String(), '2', '3', '4', '5');
 
-// $ExpectType (value?: unknown) => boolean
-bind.call(Boolean, null);
-
-// $ExpectType () => boolean
-expectType<() => boolean>(bind.apply(Boolean, [null, '1', '2', '3', '4', '5']));
+// $ExpectType (...args: string[]) => boolean
+bind.apply(Boolean, [null, '1', '2', '3', '4', '5']);
 
 // Class compatibility:
 class Foo {
     constructor(public string: string, public number: number) {}
-    declare static bind: typeof bind;
+    static bind: typeof bind;
 }
 
 // bind.call():
@@ -66,16 +54,16 @@ bind.call(Foo, null, string, number);
 // $ExpectType new () => Foo
 Foo.bind(null, string, number);
 
-// @ts-expect-error
+// $ExpectType new () => Foo
 bind.call(Foo, null, string, number, boolean);
 
-// @ts-expect-error
+// $ExpectType new () => Foo
 Foo.bind(null, string, number, boolean);
 
-// @ts-expect-error
+// $ExpectType new () => Foo
 bind.call(Foo, null, string, number, boolean, undefined);
 
-// @ts-expect-error
+// $ExpectType new () => Foo
 Foo.bind(null, string, number, boolean, undefined);
 
 // bind.apply():
