+++ types/function-bind/ts3.9/implementation.d.ts
