+++ types/function-bind/ts3.9/index.d.ts
